Update JobFilesPanel.js
Remove fileRecords parameter to the callback as the callback seems to infer the type of fileRecords incorrectly and the actual file data is lost. 

The callback is nested in another callback where fileRecords is in scope, so within the callback it is still available. 

This probably isn't a good fix, but fixes my problems.