added intermediate certificate to java truststore for ausspass's cert
added intermediate certificate to java truststore for ausspass's cert.  Note: this will only work if you build an image with spring-boot:build-image and run with docker.  It wont work if you're hosting the java app within your own JVM as the certificate would need to be injected into the hosts truststore.